rtl/tb: Introduce bare-bones instruction testbench
Trying to use core tb with compiler gernerated assembly was a litte
cumbersome when trying to test and implement individual instructions.
Introduce a testbench for testing individual instructions directly. This
test bench needs to be cycle accurate and tightly coupled to hw, so the
c++ eval step has been moved to the falling edge of the clock to make it
obvious when an internal signal is changed from c++ instead of rtl.